requirements.txt
numpy
pandas
matplotlib
seaborn
eventlet
Flask-SocketIO==4.3.1 
python-engineio==3.13.2 
python-socketio==4.6.0 
opencv-python
pillow
